Split out ref_ir.rs
