+++ crates/proc/src/ir.rs
@@ -4,20 +4,66 @@
 //
 // Copyright © 2018 Corporation for Digital Scholarship
 
-use crate::disamb::Nfa;
 use crate::prelude::*;
 use citeproc_io::output::markup::Markup;
 use citeproc_io::output::LocalizedQuotes;
 use csl::Atom;
-use csl::{Affixes, Choose, Element, Formatting, GivenNameDisambiguationRule, DateVariable, TextElement};
+use csl::{Affixes, Choose, Formatting, GivenNameDisambiguationRule, DateVariable, TextElement};
 use csl::{NumberVariable, StandardVariable, Variable};
-use crate::disamb::names::RefNameIR;
 
 use std::sync::Arc;
 
 pub mod transforms;
 
 pub type IrSum<O> = (IR<O>, GroupVars);
+
+// Intermediate Representation
+#[derive(Debug, Clone)]
+pub enum IR<O: OutputFormat = Markup> {
+    // no (further) disambiguation possible
+    Rendered(Option<CiteEdgeData<O>>),
+    // the name block,
+    Name(Arc<Mutex<NameIR<O>>>),
+
+    /// a single <if disambiguate="true"> being tested once means the whole <choose> is re-rendered in step 4
+    /// or <choose><if><conditions><condition>
+    /// Should also include `if variable="year-suffix"` because that could change.
+    ConditionalDisamb(Arc<Mutex<ConditionalDisambIR<O>>>),
+    YearSuffix(YearSuffix<O>),
+
+    // Think:
+    // <if disambiguate="true" ...>
+    //     <text macro="..." />
+    //     <text macro="..." />
+    //     <text variable="year-suffix" />
+    //     <text macro="..." />
+    // </if>
+    // = Seq[
+    //     Rendered(...), // collapsed multiple nodes into one rendered
+    //     YearSuffix(Explicit(Text(Variable::YearSuffix), T)),
+    //     Rendered(..)
+    // ]
+    Seq(IrSeq<O>),
+
+    /// Only exists to aggregate the counts of names
+    NameCounter(IrNameCounter<O>),
+}
+
+/// # Disambiguation and group_vars
+///
+/// IrSeq needs to hold things 
+#[derive(Default, Debug, PartialEq, Eq, Clone)]
+pub struct IrSeq<O: OutputFormat> {
+    pub contents: Vec<IrSum<O>>,
+    pub formatting: Option<Formatting>,
+    pub affixes: Option<Affixes>,
+    pub delimiter: Atom,
+    pub display: Option<DisplayMode>,
+    pub quotes: Option<LocalizedQuotes>,
+    pub text_case: TextCase,
+    /// If this is None, this sequence is simply an implicit conditional
+    pub dropped_gv: Option<GroupVars>,
+}
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum DisambPass {
@@ -55,147 +101,6 @@ pub enum YearSuffixHook {
     // Clone element into here, because we already know it's a <text variable="" />
     Explicit(TextElement),
     Plain,
-}
-
-impl Eq for RefIR {}
-impl PartialEq for RefIR {
-    fn eq(&self, _other: &Self) -> bool {
-        false
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug)]
-pub enum RefIR {
-    /// A piece of output that a cite can match in the final DFA.
-    /// e.g.
-    ///
-    /// ```txt
-    /// EdgeData::Output(r#"<span style="font-weight: bold;">"#)
-    /// EdgeData::Output("Some title, <i>23/4/1969</i>")
-    /// EdgeData::Locator
-    /// ```
-    ///
-    /// Each is interned into an `Edge` newtype referencing the salsa database.
-    Edge(Option<Edge>),
-
-    /// When constructing RefIR, we know whether the names variables exist or not.
-    /// So we don't have to handle 'substitute' any special way -- just drill down into the
-    /// names element, apply its formatting, and end up with
-    ///
-    /// ```txt
-    /// [
-    ///     Edge("<whatever formatting>"),
-    ///     // whatever the substitute element outputted
-    ///     Edge("</whatever>")
-    /// ]
-    /// ```
-    ///
-    /// The Nfa represents all the edge streams that a Names block can output for one of its
-    /// variables.
-    Name(RefNameIR, Nfa),
-
-    /// A non-string EdgeData can be surrounded by a Seq with other strings to apply its
-    /// formatting. This will use `OutputFormat::stack_preorder() / ::stack_postorder()`.
-    ///
-    /// ```txt
-    /// RefIR::Seq(vec![
-    ///     EdgeData::Output("<i>"),
-    ///     EdgeData::Locator,
-    ///     EdgeData::Output("</i>"),
-    /// ])
-    /// ```
-    Seq(RefIrSeq),
-    // Could use this to apply a FreeCond set to a reference to create a path through the
-    // constructed NFA.
-    // See the module level documentation for `disamb`.
-    // Branch(Arc<Conditions>, Box<IR<O>>),
-}
-
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-pub struct RefIrSeq {
-    pub contents: Vec<RefIR>,
-    pub formatting: Option<Formatting>,
-    pub affixes: Option<Affixes>,
-    pub delimiter: Atom,
-    pub quotes: Option<LocalizedQuotes>,
-    pub text_case: TextCase,
-}
-
-impl RefIR {
-    pub fn debug(&self, db: &dyn IrDatabase) -> String {
-        match self {
-            RefIR::Edge(Some(e)) => format!("{:?}", db.lookup_edge(*e)),
-            RefIR::Edge(None) => "None".into(),
-            RefIR::Seq(seq) => {
-                let mut s = String::new();
-                s.push_str("[");
-                let mut seen = false;
-                for x in &seq.contents {
-                    if seen {
-                        s.push_str(",");
-                    }
-                    seen = true;
-                    s.push_str(&x.debug(db));
-                }
-                s.push_str("]");
-                s
-            }
-            RefIR::Name(rnir, _nfa) => format!("NameVariable::{:?}", rnir.variable),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            RefIR::Edge(None) => true,
-            RefIR::Seq(seq) => seq.contents.is_empty(),
-            RefIR::Name(_rnir, nfa) => nfa.is_empty(),
-            _ => false,
-        }
-    }
-
-    pub(crate) fn keep_first_ysh(&mut self, ysh_explicit_edge: Edge, ysh_plain_edge: Edge, ysh_edge: Edge) {
-        let found = &mut false;
-        self.visit_ysh(ysh_explicit_edge, &mut |opt_e| {
-            if !*found {
-                // first time
-                *found = true;
-                *opt_e = Some(ysh_edge);
-            } else {
-                // subsequent ones are extraneous, so make them disappear
-                *opt_e = None;
-            }
-            false
-        });
-        self.visit_ysh(ysh_plain_edge, &mut |opt_e| {
-            if !*found {
-                *found = true;
-                *opt_e = Some(ysh_edge);
-            } else {
-                *opt_e = None;
-            }
-            false
-        });
-    }
-
-    pub(crate) fn visit_ysh<F>(&mut self, ysh_edge: Edge, callback: &mut F) -> bool
-    where
-        F: (FnMut(&mut Option<Edge>) -> bool),
-    {
-        match self {
-            RefIR::Edge(ref mut opt_e) if *opt_e == Some(ysh_edge) => callback(opt_e),
-            RefIR::Seq(seq) => {
-                for ir in seq.contents.iter_mut() {
-                    let done = ir.visit_ysh(ysh_edge, callback);
-                    if done {
-                        return true;
-                    }
-                }
-                false
-            }
-            _ => false,
-        }
-    }
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -259,38 +164,6 @@ pub struct ConditionalDisambIR<O: OutputFormat> {
     pub ir: Box<IR<O>>,
 }
 
-// Intermediate Representation
-#[derive(Debug, Clone)]
-pub enum IR<O: OutputFormat = Markup> {
-    // no (further) disambiguation possible
-    Rendered(Option<CiteEdgeData<O>>),
-    // the name block,
-    Name(Arc<Mutex<NameIR<O>>>),
-
-    /// a single <if disambiguate="true"> being tested once means the whole <choose> is re-rendered in step 4
-    /// or <choose><if><conditions><condition>
-    /// Should also include `if variable="year-suffix"` because that could change.
-    ConditionalDisamb(Arc<Mutex<ConditionalDisambIR<O>>>),
-    YearSuffix(YearSuffix<O>),
-
-    // Think:
-    // <if disambiguate="true" ...>
-    //     <text macro="..." />
-    //     <text macro="..." />
-    //     <text variable="year-suffix" />
-    //     <text macro="..." />
-    // </if>
-    // = Seq[
-    //     Rendered(...), // collapsed multiple nodes into one rendered
-    //     YearSuffix(Explicit(Text(Variable::YearSuffix), T)),
-    //     Rendered(..)
-    // ]
-    Seq(IrSeq<O>),
-
-    /// Only exists to aggregate the counts of names
-    NameCounter(IrNameCounter<O>),
-}
-
 #[derive(Debug, Clone)]
 pub struct IrNameCounter<O: OutputFormat> {
     pub name_irs: Vec<NameIR<O>>,
@@ -321,23 +194,6 @@ impl<O: OutputFormat> IrNameCounter<O> {
         (IR::Rendered(Some(CiteEdgeData::Output(built))), GroupVars::Important)
     }
 }
-
-// #[derive(Debug, Clone)]
-// pub struct RefIrNameCounter {
-//     name_irs: Vec<RefNameIR>,
-// }
-// impl RefIrNameCounter {
-//     fn count(&self) -> u32 {
-//         500
-//     }
-//     pub fn render_ref(&self, db: &dyn IrDatabase, ctx: &RefContext<'_, Markup>, stack: Formatting, piq: Option<bool>) -> (RefIR, GroupVars) {
-//         let count = self.count();
-//         let fmt = ctx.format;
-//         let out = fmt.output_in_context(fmt.text_node(format!("{}", count), None), stack, piq);
-//         let edge = db.edge(EdgeData::<Markup>::Output(out));
-//         (RefIR::Edge(Some(edge)), GroupVars::Important)
-//     }
-// }
 
 impl<O> IR<O>
 where
@@ -393,12 +249,6 @@ where
 impl<O: OutputFormat> Default for IR<O> {
     fn default() -> Self {
         IR::Rendered(None)
-    }
-}
-
-impl Default for RefIR {
-    fn default() -> Self {
-        RefIR::Edge(None)
     }
 }
 
@@ -528,23 +378,6 @@ impl IR<Markup> {
 //         }
 //     }
 // }
-
-/// # Disambiguation and group_vars
-///
-/// IrSeq needs to hold things 
-#[derive(Default, Debug, PartialEq, Eq, Clone)]
-pub struct IrSeq<O: OutputFormat> {
-    pub contents: Vec<IrSum<O>>,
-    pub formatting: Option<Formatting>,
-    pub affixes: Option<Affixes>,
-    pub delimiter: Atom,
-    pub display: Option<DisplayMode>,
-    pub quotes: Option<LocalizedQuotes>,
-    pub text_case: TextCase,
-    /// If this is None, this sequence is simply an implicit conditional
-    pub dropped_gv: Option<GroupVars>,
-}
-
 
 impl<O: OutputFormat> IR<O> {
     pub(crate) fn recompute_group_vars(&mut self) {
